Added place holders for remaining input types
